techDebt(registrationRefactor): Move creation of establishment into separate createEstablishment file
